ajustes em fornada e implementação da controller dos ingredientes

+++ Atividade_6_Pizzaria/src/main/java/br/com/soulcode/pizzaria/controllers/FornadaController.java
@@ -36,7 +36,7 @@ public class FornadaController {
     }
 
     @PutMapping(value = "/editar-fornada/{id}")
-    public ResponseEntity<Fornada> updateById(@RequestBody Fornada fornada, @PathVariable long id) {
+    public ResponseEntity<Fornada> updateById(@RequestBody Fornada fornada, @PathVariable int id) {
         return fornadaService.updateById(fornada, id);
     }
